File system structure refactored
