+++ src/index.ts
@@ -1,7 +1,7 @@
 import { API } from 'homebridge';
 
 import { PLATFORM_NAME } from './settings';
-import { WledWsHomebridgePlatform } from './platform';
+import { WledWsHomebridgePlatform } from './WledWsPlatform';
 
 /**
  * This method registers the platform with Homebridge
